+++ angular-app/src/app/components/add-offer/add-offer.component.html
@@ -13,23 +13,17 @@
             <div *ngIf="f.description.errors!.required">Description is required.</div>
         </div>
     </div>
-
     <div class="form-group">
-        <mat-form-field appearance="fill">
-            <mat-label>Choose category</mat-label>
-            <mat-select formControlName="category" (selectionChange)="onChange($event)">
-                <mat-option *ngFor='let c of categories' [value]="c.id">{{c.name}}</mat-option>
-            </mat-select>
-        </mat-form-field>
+        <select #mySelect class="browser-default custom-select" formControlName="category" (change)="onChange(mySelect.value)">
+            <option value="" disabled selected>Choose category</option>
+            <option *ngFor='let c of categories' [value]="c.id">{{c.name}}</option>
+        </select>
     </div>
-
     <div class="form-group">
-        <mat-form-field appearance="fill">
-            <mat-label>Choose type</mat-label>
-            <mat-select formControlName="type" >
-                <mat-option *ngFor='let c of types' [value]="c.id">{{c.name}}</mat-option>
-            </mat-select>
-        </mat-form-field>
+        <select class="browser-default custom-select" formControlName="type">
+            <option value="" disabled selected>Choose type</option>
+            <option *ngFor='let c of types' [value]="c.id">{{c.name}}</option>
+        </select>
     </div>
     <div class="form-group">
         <label for="place">Place:</label>
