+++ angular-app/src/app/model/category-model.ts
