+++ angular-app/src/app/model/offer-mode.ts
