+++ angular-app/src/app/model/type-model.ts
