+++ angular-app/src/app/services/type.service.spec.ts
